Add LimitBytes and SinceSeconds to `cluster-info dump`
The `kubectl cluster-info dump` command has no limits on the logs that are
gathered; most of the time only the most recent logs are important, and
gathering all of the logs can make large demands on time and space.  These
changes allow for space limits (although from the oldest of the logs) and time
back in the logs from the most recent backwards in time.